Moved the Counter class items to the separate counter class and removed it from BoardManager.  Also included the Counter classes namespace
