+++ TeamNameNotFoundException/Assets/Scripts/BoardManager.cs
@@ -2,21 +2,10 @@
 using System;
 using System.Collections.Generic; 		
 using Random = UnityEngine.Random; 		
+using Counter;
 
 public class BoardManager : MonoBehaviour
-{
-	[Serializable]
-	public class Count
-	{
-		public int minimum; 			
-		public int maximum; 	
-		public Count (int min, int max)
-		{
-			minimum = min;
-			maximum = max;
-		}
-	}
-		
+{		
 	public int columns = 8; 										
 	public int rows = 8;											
 	public Count wallCount = new Count (5, 9);
