+++ TeamNameNotFoundException/UnityVS.TeamNameNotFoundException.CSharp.csproj
@@ -70,6 +70,7 @@
     <Compile Include="Assets\DeleteScript.cs" />
     <Compile Include="Assets\NewBehaviourScript.cs" />
     <Compile Include="Assets\Scripts\BoardManager.cs" />
+    <Compile Include="Assets\Scripts\Count.cs" />
     <Compile Include="Assets\Scripts\Enemy.cs" />
     <Compile Include="Assets\Scripts\GameManager.cs" />
     <Compile Include="Assets\Scripts\Loader.cs" />
